NUEVO-F-main/MiTiendaWeb/chatbot/requirements.txt
python-telegram-bot==20.7
openai==1.3.0
pymongo==4.6.1
python-dotenv==1.0.0
logging==0.4.9.6 